% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placer-place.R
\name{clear_bet_slip}
\alias{clear_bet_slip}
\title{Clear the bet slip by clicking "Hreinsa raddir"}
\usage{
clear_bet_slip(session)
}
\arguments{
\item{session}{ChromoteSession (authenticated)}
}
\value{
Invisible logical: TRUE if the slip was cleared
}
\description{
Clear the bet slip by clicking "Hreinsa raddir"
}
