% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placer-navigate.R
\name{competition_url}
\alias{competition_url}
\title{Build URL for a competition listing page}
\usage{
competition_url(sport_id, competition_id)
}
\arguments{
\item{sport_id}{Lengjan sport ID (integer)}

\item{competition_id}{Lengjan competition ID (integer)}
}
\value{
A length-1 character URL
}
\description{
Build URL for a competition listing page
}
