% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placer-navigate.R
\name{expand_all}
\alias{expand_all}
\title{Click "Sjá allt" button to expand truncated match list (via JS)}
\usage{
expand_all(session)
}
\arguments{
\item{session}{A ChromoteSession (authenticated)}
}
\value{
invisible(NULL)
}
\description{
Click "Sjá allt" button to expand truncated match list (via JS)
}
