% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisko.R
\name{gisko_marginals_from_pbar}
\alias{gisko_marginals_from_pbar}
\title{Marginal summaries of an integrated scoreline PMF}
\usage{
gisko_marginals_from_pbar(pbar)
}
\arguments{
\item{pbar}{`n x n` matrix, rows = home, cols = away.}
}
\value{
list `p_outcome` (named home/draw/away), `home_pmf`, `away_pmf`
  (named by goals), `gd_pmf` (named by signed goal difference).
}
\description{
Marginal summaries of an integrated scoreline PMF
}
