% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placer-navigate.R
\name{match_url}
\alias{match_url}
\title{Build URL for a match detail page}
\usage{
match_url(match_id, sport_id)
}
\arguments{
\item{match_id}{Lengjan match ID (character or integer)}

\item{sport_id}{Lengjan sport ID (integer)}
}
\value{
A length-1 character URL
}
\description{
Build URL for a match detail page
}
