% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_boxscore.R
\name{nba.GameBoxScoreAdvanced}
\alias{nba.GameBoxScoreAdvanced}
\title{Box Score - Advanced}
\usage{
nba.GameBoxScoreAdvanced(gameid, info_type = "team", ...)
}
\arguments{
\item{gameid}{String}

\item{info_type}{String}

\item{...}{List}
}
\value{
dataframe
}
\description{
Box Score - Advanced
}
\examples{
df <- nba.GameBoxScoreAdvanced('0021601219')
names(df)
head(df, 3)
}
