% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_boxscore.R
\name{nba.GameBoxScoreFourFactors}
\alias{nba.GameBoxScoreFourFactors}
\title{Box Score - Four Factors}
\usage{
nba.GameBoxScoreFourFactors(gameid, info_type = "team", ...)
}
\arguments{
\item{gameid}{String}

\item{info_type}{String}

\item{...}{List}
}
\value{
dataframe
}
\description{
Box Score - Four Factors
}
\examples{
df <- nba.GameBoxScoreFourFactors('0021601219')
names(df)
head(df, 3)
}
