% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_boxscore.R
\name{nba.GameBoxScoreHustle}
\alias{nba.GameBoxScoreHustle}
\title{Box Score - Hustle}
\usage{
nba.GameBoxScoreHustle(gameid, info_type = "team", ...)
}
\arguments{
\item{gameid}{String}

\item{info_type}{String}

\item{...}{List}
}
\value{
dataframe
#@export
}
\description{
Box Score - Hustle
}
\examples{
\dontrun{
df <- nba.GameBoxScoreHustle('0021601219')
names(df)
head(df, 3)
}
}
