% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_boxscore.R
\name{nba.GameBoxScoreScoring}
\alias{nba.GameBoxScoreScoring}
\title{Box Score - Scoring}
\usage{
nba.GameBoxScoreScoring(gameid, info_type = "team", ...)
}
\arguments{
\item{gameid}{String}

\item{info_type}{String}

\item{...}{List}
}
\value{
dataframe
}
\description{
Box Score - Scoring
}
\examples{
df <- nba.GameBoxScoreScoring('0021601219')
names(df)
head(df, 3)
}
