% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_boxscore.R
\name{nba.GameBoxScoreSummary}
\alias{nba.GameBoxScoreSummary}
\title{Box Score - Summary}
\usage{
nba.GameBoxScoreSummary(gameid, info_type = "summary", ...)
}
\arguments{
\item{gameid}{String}

\item{info_type}{String}

\item{...}{List}
}
\value{
dataframe
}
\description{
Box Score - Summary
}
\examples{
df <- nba.GameBoxScoreSummary('0021601219', info_type = 'linescore')
names(df)
head(df, 3)
}
