% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_boxscore.R
\name{nba.GameBoxScoreTracking}
\alias{nba.GameBoxScoreTracking}
\title{Box Score - Tracking}
\usage{
nba.GameBoxScoreTracking(gameid, info_type = "team", ...)
}
\arguments{
\item{gameid}{String}

\item{info_type}{String}

\item{...}{List}
}
\value{
dataframe
}
\description{
Box Score - Tracking
}
\examples{
df <- nba.GameBoxScoreTracking('0021601219')
names(df)
head(df, 3)
}
