% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_boxscore.R
\name{nba.GameBoxScoreTraditional}
\alias{nba.GameBoxScoreTraditional}
\title{Box Score - Traditional}
\usage{
nba.GameBoxScoreTraditional(gameid, info_type = "team", ...)
}
\arguments{
\item{gameid}{String}

\item{info_type}{String}

\item{...}{List}
}
\value{
dataframe
}
\description{
Box Score - Traditional
}
\examples{
df <- nba.GameBoxScoreTraditional('0021601219')
names(df)
head(df, 3)
}
