% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_boxscore.R
\name{nba.GameBoxScoreUsage}
\alias{nba.GameBoxScoreUsage}
\title{Box Score - Usage}
\usage{
nba.GameBoxScoreUsage(gameid, info_type = "team", ...)
}
\arguments{
\item{gameid}{String}

\item{info_type}{String}

\item{...}{List}
}
\value{
dataframe
}
\description{
Box Score - Usage
}
\examples{
df <- nba.GameBoxScoreUsage('0021601219')
names(df)
head(df, 3)
}
