% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_game_logs.R
\name{nba.LeagueGameLog}
\alias{nba.LeagueGameLog}
\title{Get Game Logs}
\usage{
nba.LeagueGameLog(...)
}
\arguments{
\item{...}{}
}
\value{
dataframe
}
\description{
Get Game Logs
}
\examples{
df <- nba.LeagueGameLog()
names(df)
head(df)
}
