% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{read_table}
\alias{read_table}
\title{Read a table back as a tibble.}
\usage{
read_table(table, root = here::here("data"), filter = list())
}
\arguments{
\item{table}{one of names(schemas()).}

\item{root}{filesystem root.}

\item{filter}{optional named list of column=value filters pushed down to Arrow.}
}
\value{
tibble
}
\description{
Read a table back as a tibble.
}
