% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{validate_against_schema}
\alias{validate_against_schema}
\title{Validate a data frame against a schema.
Raises a diagnostic error on the first problem found.}
\usage{
validate_against_schema(df, table)
}
\description{
Validate a data frame against a schema.
Raises a diagnostic error on the first problem found.
}
